# Random tests optimization (take 2)