--- a/appveyor.yml
+++ b/appveyor.yml
@@ -31,7 +31,8 @@ install:
 
     # Install specified version of numpy and dependencies
     - "conda install --yes -c conda-forge numpy scipy nose pytest setuptools ipython git unyt
-    Cython sympy fastcache h5py matplotlib=3.1.3 mock pandas cartopy conda-build pooch pyyaml"
+    Cython sympy fastcache h5py matplotlib=3.1.3 mock pandas cartopy conda-build pooch pyyaml
+    nose-timer pykdtree"
     # install yt
     - "pip install -e ."
 
@@ -39,7 +40,7 @@ install:
 build: false
 
 test_script:
-    - "nosetests --nologcapture -sv --traverse-namespace yt"
+    - "nosetests --with-timer --timer-top-n=20 --nologcapture --with-xunit -sv --traverse-namespace yt"
 
 # Enable this to be able to login to the build worker. You can use the
 # `remmina` program in Ubuntu, use the login information that the line below