--- a/tests/test_minimal_requirements.txt
+++ b/tests/test_minimal_requirements.txt
@@ -9,3 +9,4 @@ pyyaml>=4.2b1
 coverage==4.5.1
 codecov==2.0.15
 unyt==2.7.2
+mock