--- a/tests/test_requirements.txt
+++ b/tests/test_requirements.txt
@@ -6,7 +6,7 @@ glueviz==0.13.3
 h5py==2.10.0
 ipython==7.1.1
 matplotlib==3.3.0
-mock==2.0.0; python_version < '3.0'
+mock
 nose-timer==1.0.0
 nose==1.3.7
 pandas==0.25.3