--- a/tests/test_requirements.txt
+++ b/tests/test_requirements.txt
@@ -28,3 +28,4 @@ firefly_api>=0.0.2
 f90nml>=1.1.2
 MiniballCpp>=0.2.1
 pooch>=0.7.0
+pykdtree==1.3.1