Currently 10 unit tests account for **25%** of build time on Travis (and I assume appveyor):

```
[success] 5.40% yt.data_objects.tests.test_projection.test_projection: 82.3893s
[success] 4.90% yt.visualization.tests.test_particle_plot.TestParticlePhasePlotSave.test_particle_phase_plot: 74.7280s
[success] 3.55% yt.visualization.tests.test_geo_projections.TestGeoProjections.test_geo_projections: 54.0847s
[success] 2.49% yt.frontends.gadget.tests.test_outputs.test_gadget_binary: 37.9936s
[success] 2.21% yt.visualization.tests.test_particle_plot.TestParticleProjectionPlotSave.test_particle_plot_c: 33.6426s
[success] 1.80% yt.data_objects.tests.test_slice.test_slice: 27.3991s
[success] 1.62% yt.visualization.volume_rendering.tests.test_scene.RotationTest.test_rotation: 24.6765s
[success] 1.14% yt.frontends.stream.tests.test_stream_particles.test_stream_particles: 17.4531s
[success] 1.10% yt.visualization.tests.test_callbacks.test_mesh_lines_callback: 16.7762s
[success] 1.08% yt.visualization.tests.test_plotwindow.TestPlotWindowSave.test_projection_plot_c: 16.3965s
```

This PR addresses some of them by:
 1) mocking MPL's `print_figure` method, as we're not actually analyzing images in any way
 2) using `pykdtree` instead of `scipy.kdtree` for `Cartopy`. On my laptop it yields 3x speed up for *test_geo_projections*