@neutrinoceros 2020-07-22
Well done, you won close to 20min on the Appveyor run !